Apply the environments in preFlightChecks, so that tools provided by the dependencies can be detected.
This allows to require tools (like qmake) in plugins that are not in
the path by default. it is also a nice self-test for the configured
environments. Corollary - you do not want the qmake of a platform
provided Qt in the path, because that would provide a fallback that
might hide errors in the setup.