More preparations for branch building.
Add a property to SourceCodeProvider that switches parsing of commits
in branches on and off (off by default). Add base class methods
getRevisionsSinceAllBranches and getCurrentRevisionAllBranches that
will be called instead of getRevisionsSince and getCurrentRevision if
branch commit parsing is enabled. These methods now need to be
implemented in the different SCM implementations.