Finish implementation of SettingResolver, add configurable way to add subdirectories to uploaders.
This probably needs cleanup. There are a few new settings that need to
be verified on the existing build machines. Updated the tests to work
with the revision infos being cached now. Make DirectoryTreeCopyAction
work with StringResolvers, introducing a new common subclass of
PathResolver and SettingResolver. Both also now understand an option
to pass in patterns to be applied during evaluation.