Introduce the report and notify phases.
This makes it explicit at what point in time the reports and notifications should be created, instead of hooking those into the shutdown phase (which is only for cleanup purposes). Implement conditional deletion of the log directory structure during shutdown. Deletion of the log directory is enabled by default, and will be disabled automatically if an exception occurs. Fix arguments for debug call. Be more precise - 0 is not None, and return code 0 is not an error.